Update gql invoice stats and remove domain object
