+++ graphql/src/loader/invoice_line.rs
@@ -1,4 +1,3 @@
-use domain::invoice::InvoicePricing;
 use domain::invoice_line::{InvoiceLine, InvoiceLineFilter};
 use domain::EqualFilter;
 use repository::{InvoiceLineRepository, RepositoryError, StorageConnectionManager};
@@ -32,24 +31,5 @@ impl Loader<String> for InvoiceLineLoader {
             .collect();
 
         Ok(result)
-    }
-}
-
-pub struct InvoiceLineStatsLoader {
-    pub connection_manager: StorageConnectionManager,
-}
-
-#[async_trait::async_trait]
-impl Loader<String> for InvoiceLineStatsLoader {
-    type Value = InvoicePricing;
-    type Error = RepositoryError;
-
-    async fn load(
-        &self,
-        invoice_ids: &[String],
-    ) -> Result<HashMap<String, Self::Value>, Self::Error> {
-        let connection = self.connection_manager.connection()?;
-        let repo = InvoiceLineRepository::new(&connection);
-        Ok(repo.stats(invoice_ids)?)
     }
 }
